+++ Annex/src/utopia/annex/model/schrodinger/PutSchrodinger.scala
@@ -40,4 +40,25 @@ trait PutSchrodinger[S, I] extends Schrodinger[Try[I], S]
 			}
 		case None => localModified
 	}
+	
+	
+	// OTHER    -------------------------
+	
+	/**
+	  * Successfully completes this schrödinger instance
+	  * @param successResult Successful result
+	  */
+	def succeedWith(successResult: I) = complete(Success(successResult))
+	
+	/**
+	  * Completes this Sscrödinger with a failure result
+	  * @param failureResult A failure result
+	  */
+	def failWith(failureResult: Failure[I]) = complete(failureResult)
+	
+	/**
+	  * Completes this Sscrödinger with a failure result
+	  * @param error Error assigned as failure
+	  */
+	def failWith(error: Throwable): Unit = failWith(Failure(error))
 }